Move data transformation out to lib
